IMPALA-3772: Fix admission control flaky stress test
The admission control stress test could occasionally fail
with some queued queries timing out. At some point the
default statestore topic update rpc frequency was increased
to 2sec, and because the updates trigger the admission
control queue check (to determine if queued queries can be
dequeued), it is possible that the topic updates weren't
frequent enough for a large admission control queue to fully
drain within the 60sec queue timeout.

In the test that failed, queries were submitted round-robin
so the admission control behavior is non-deterministic.
There isn't enough information to prove that this is
definitely the issue, but it seems likely and improving the
test seems valuable regardless.

This change reduces the statestore topic update frequency to
500ms so that state can be shared faster and the queued
requests have more chances to drain within the 60sec queue
timeout. The code was also a bit confusing before because it
was waiting for statestore topic updates but the test had
only changed the heartbeat frequency (which is a different
RPC). This now sets the lower frequency for both and makes
the code a bit more clear.

Change-Id: I235a14c4674240dc0a01dabb664da87c752153cf
Reviewed-on: http://gerrit.cloudera.org:8080/3450
Reviewed-by: Matthew Jacobs <mj@cloudera.com>
Tested-by: Matthew Jacobs <mj@cloudera.com>