IMPALA-3800: Python 2.6 support for collect_minidumps.py
Python 2.6 is the default python version shipped with CentOS 6.6 and
the minidump collection script needs to run there, too. However, the
tarfile library version shipped with python 2.6 does not support using
TarFile objects as context managers. This change wraps the usage of
tarfile.TarFile in contextlib.closing to support python 2.6, too.

Testing: I verified that this script runs with python 2.7 on testdata
generated using the generate_minidump_collection_testdata.py script. I
also copied an updated version to a CentOS 6.6 cluster and verified that
the problem is resolved there.

Change-Id: Ic9028626fa829ff7571d2a731c2fcd7e15e2ce36
Reviewed-on: http://gerrit.cloudera.org:8080/3525
Reviewed-by: Michael Brown <mikeb@cloudera.com>
Reviewed-by: Taras Bobrovytsky <tbobrovytsky@cloudera.com>
Reviewed-by: Dan Hecht <dhecht@cloudera.com>
Tested-by: Lars Volker <lv@cloudera.com>