IMPALA-3817: Ensure filter hash function is the same on all hardware.
If some nodes in a cluster support the hardware CRC instruction, then
those nodes would use that hash function when testing filters. This
change uses Murmur2 always. Murmur2 is slower, but a perf run on a
16-node cluster showed negligible performance differences.

This impacts DataStreamMgr's hash function, but that that also doesn't
show a perf difference.

(Also, remove unused hash function from HashUtil that was made dead
code by an earlier commit.)

Change-Id: Ia09b67f9e987af3e2c8ac12c347b95a7e09ce6fa
Reviewed-on: http://gerrit.cloudera.org:8080/3566
Reviewed-by: Jim Apple <jbapple@cloudera.com>
Tested-by: Internal Jenkins