IMPALA-3756: Fix wrong argument type in HiveStringsTest
One of the tests in HiveStringsTest passed Text as input argument
to TestHiveUdf() instead of String. It worked fine if the argument
of UDFLength.evaluate() is of type Text but that doesn't seem to be
the case in code coverage build. The fix is to call TestHiveUdf()
with String and lets it cast the argument to the right type based
on the signature of the UDF method. A test-only UdfExecutor ctor
is also removed so we can exercise the actual ctor used by JNI calls
from the BE.

Change-Id: I662e6286dac601ae0e45f18545ef149724aa047e
Reviewed-on: http://gerrit.cloudera.org:8080/3617
Reviewed-by: Michael Ho <kwho@cloudera.com>
Tested-by: Internal Jenkins