IMPALA-3090: always log memory limit errors
Consistently log memory limit errors so that the error message contains
a dump of the query memory trackers at the time that the memory limit
was hit (instead of after the fact when the query is already partially
cleaned up).

Testing:
Exhaustive build passed. Ran local stress test for a bit.

Change-Id: If5ec5572b0e26898da352b7e6b11eb01c6edb2e5
Reviewed-on: http://gerrit.cloudera.org:8080/4049
Reviewed-by: Tim Armstrong <tarmstrong@cloudera.com>
Tested-by: Internal Jenkins