IMPALA-4004: Don't access nested types in test_failpoints.py
As part of fixing IMPALA-3692, the query in test_failpoints.py was
updated to have a predicate on a string column in a parquet table.
The update to the query was based on the failing query in the bug
which happens to access nested columns. Apparently, this doesn't
quite work with the legacy join and agg.

This change fixes the test to also work with legacy join and agg.
With the debug actions added in the fix of IMPALA-3692, it's not
necessary to access nested types to reproduce the problem as long
as there is a predicate on a string column.

Change-Id: Idc5e67b9748a13fcd76ea5fe140e2e6b18e809b7
Reviewed-on: http://gerrit.cloudera.org:8080/4074
Reviewed-by: Alex Behm <alex.behm@cloudera.com>
Tested-by: Internal Jenkins
(cherry picked from commit b251fb0615d2bcdaade93a5d939c7ab6bc64960c)