IMPALA-3986: Python hive-client may fail silently while dropping partitions
It was noticed during a test run that a partition existed in the HMS
when it shouldn't have. The only possible explanation for that is that
the python based hive client failed to drop the partition when asked
to and we did not check for an error when it did so.

This patch just adds an assert to make sure that the partition is
dropped when instructed to, else it will cause the test to fail in
the right place.

Change-Id: I4241225d19c75a79179a61fef0c75c06a9a2ac39
Reviewed-on: http://gerrit.cloudera.org:8080/4083
Reviewed-by: Matthew Jacobs <mj@cloudera.com>
Tested-by: Internal Jenkins