IMPALA-3944: Fix crash in SimpleScheduler
SimpleScheduler::BackendConfig maintains two maps, hostname->IP and
IP->[list of backends]. In situations where multiple backends run on a
single host, BackendConfig::RemoveBackend was erroneously removing the
hostname->IP mapping, even when the list of backends for a host was not
empty. The fix is to only remove this mapping when the last backend of a
host gets removed.

I will push a test in a separate change.

Change-Id: I1c2c58c2f9e7de0936aba03dcce86cdc4b31a866
Reviewed-on: http://gerrit.cloudera.org:8080/4102
Reviewed-by: Lars Volker <lv@cloudera.com>
Reviewed-by: Matthew Jacobs <mj@cloudera.com>
Tested-by: Internal Jenkins