IMPALA-3979: Fix dynamic linking for Impala
When building Impala, the user can choose between two different
build modes: static and dynamically linked libraries. The first
one is used for release builds, while the latter one is used
mainly locally for developers to shorten build times.

However, for some time now, the shared build has been broken in
some configurations. This patch fixes this issue. The main
solution is to move the webserver into the Util library and to
remove the manual linking of Util into the error-utils-test.

Change-Id: I2fb1efd600bfaf17de730186fc80a0dbd976a049
Reviewed-on: http://gerrit.cloudera.org:8080/4108
Tested-by: Internal Jenkins
Reviewed-by: Tim Armstrong <tarmstrong@cloudera.com>