IMPALA-4019: initialize member variables in HdfsTableSink
This fixes a bug that could potentially lead to the IMPALA-4019 crash.
default_partition_ should be initialized to NULL so that we can check
if the HdfsTableSink was correctly initialized with a default
partition in Open().

After this change the query will fail with an error status instead of
potentially crashing if table_desc_->partition_descriptors() is missing
the default partition.

Change-Id: If62def81f7d06b42ffbd0dc2dcb0e777421a1edf
Reviewed-on: http://gerrit.cloudera.org:8080/4171
Reviewed-by: Alex Behm <alex.behm@cloudera.com>
Reviewed-by: Dan Hecht <dhecht@cloudera.com>
Tested-by: Internal Jenkins