IMPALA-4102: Remote Kudu reads should be reported
Adds a profile counter for the number of kudu scan tokens
(ranges) that are "expected" to be remote.

Testing: Manual; Have been running with this on the Kudu
cluster. Cannot easily simulate this in the minicluster
because the scheduler considers multiple impalads on the
same host to be local for the purposes of determining
locality. See BackendConfig::LookUpBackendIp().

Change-Id: I74fd5773c4ae10267de80b6572d93197a4131696
Reviewed-on: http://gerrit.cloudera.org:8080/4687
Reviewed-by: Alex Behm <alex.behm@cloudera.com>
Tested-by: Internal Jenkins