IMPALA-4099: Fix the error message while loading UDFs with no JARs
Currently we throw a generic exception due to a missing check in
the code for cases where the UDF has no resources associated with
it. This is not super useful from supportability point of view.
This commit fixes that. This also cleans up the code a little
by moving these checks out of the extractFunctions() method.

Change-Id: I167426ea96b0a41374227ef238b6f60e4184a9d7
Reviewed-on: http://gerrit.cloudera.org:8080/4365
Reviewed-by: Alex Behm <alex.behm@cloudera.com>
Reviewed-by: Marcel Kornacker <marcel@cloudera.com>
Tested-by: Internal Jenkins
(cherry picked from commit 94fc6a3d07aaf19591b832d4fc88eee831db9f99)
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/16774
Reviewed-by: Bharath Vissapragada <bharathv@cloudera.com>
Tested-by: Jeges Attila <attilaj@cloudera.com>